Run rename script
