+++ environment.yml
@@ -1,4 +1,4 @@
-name: stactools-ephemeral
+name: stactools-usgs-nlcd
 channels:
   - conda-forge
   - defaults
