+++ tests/test_commands.py
@@ -4,12 +4,12 @@ from tempfile import TemporaryDirectory
 import pystac
 from stactools.testing import CliTestCase
 
-from stactools.ephemeral.commands import create_ephemeralcmd_command
+from stactools.usgs_nlcd.commands import create_usgsnlcd_command
 
 
 class CommandsTest(CliTestCase):
     def create_subcommand_functions(self):
-        return [create_ephemeralcmd_command]
+        return [create_usgsnlcd_command]
 
     def test_create_collection(self):
         with TemporaryDirectory() as tmp_dir:
@@ -19,7 +19,7 @@ class CommandsTest(CliTestCase):
             destination = os.path.join(tmp_dir, "collection.json")
 
             result = self.run_command(
-                ["ephemeralcmd", "create-collection", destination])
+                ["usgsnlcd", "create-collection", destination])
 
             self.assertEqual(result.exit_code,
                              0,
@@ -41,7 +41,7 @@ class CommandsTest(CliTestCase):
             # Example:
             destination = os.path.join(tmp_dir, "collection.json")
             result = self.run_command([
-                "ephemeralcmd",
+                "usgsnlcd",
                 "create-item",
                 "/path/to/asset.tif",
                 destination,
